requirements.txt
streamlit
python-docx
python-dotenv
langchain
langchain-community
openai
chromadb
